+++ src/dependencies/lcommon/SerializeBuffer.h
@@ -67,8 +67,8 @@
 #   define le64toh(x) __bswap_64 (x)
 #endif
 #endif
-const int MAX_ALLOC_SIZE = 32 * 1024 * 1024 * 1024; //128MB
-const int MAX_BUFFER_SIZE = 128 * 1024 * 1024; //128kb
+const long long int MAX_ALLOC_SIZE = 32 * 1024 * 1024 * 1024; //128MB
+const long long int MAX_BUFFER_SIZE = 128 * 1024 * 1024; //128kb
 
 typedef void (*buffer_flushf)(void* context, const char* data, size_t size);
 typedef void (*buffer_fillf)(void* context, std::vector<char>& buffer,
